Update
Uploading code with link to database. Not working with a few variables.
Will review. Need to work on link from Nestor's pages to pass in
variable

+++ OrganDonorSystem/OrganDonorSystem/ViewModels/DonorInfoIndexViewModel.cs
@@ -9,7 +9,7 @@ namespace OrganDonorSystem.ViewModels
     {
 
         public int donorID { get; set; }
-        public int age { get; set; }
+        public long age { get; set; }
         public string gender { get; set; }
         public string phone { get; set; }
         public DateTime registration { get; set; }
